+++ src/AnomalyDetectors/IsolationForest.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\AnomalyDetectors;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Scoring;
 use Rubix\ML\Ranking;
 use Rubix\ML\DataType;
 use Rubix\ML\Estimator;
@@ -46,7 +47,7 @@ use const Rubix\ML\EPSILON;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class IsolationForest implements Estimator, Learner, Ranking, Persistable, Stringable
+class IsolationForest implements Estimator, Learner, Scoring, Ranking, Persistable, Stringable
 {
     use PredictsSingle, ScoresSingle;