+++ src/Scoring.php
@@ -4,28 +4,21 @@ namespace Rubix\ML;
 
 use Rubix\ML\Datasets\Dataset;
 
-/**
- * @deprecated
- */
-interface Ranking extends Estimator
+interface Scoring extends Estimator
 {
     /**
      * Return the anomaly scores assigned to the samples in a dataset.
      *
-     * @deprecated
-     *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @return float[]
      */
-    public function rank(Dataset $dataset) : array;
+    public function score(Dataset $dataset) : array;
 
     /**
-     * Return the score given to a single sample.
-     *
-     * @deprecated
+     * Return the anomaly score given to a single sample.
      *
      * @param (string|int|float)[] $sample
      * @return float
      */
-    public function rankSample(array $sample) : float;
+    public function scoreSample(array $sample) : float;
 }