+++ tests/AnomalyDetectors/LocalOutlierFactorTest.php
@@ -3,7 +3,7 @@
 namespace Rubix\ML\Tests\AnomalyDetectors;
 
 use Rubix\ML\Learner;
-use Rubix\ML\Ranking;
+use Rubix\ML\Scoring;
 use Rubix\ML\DataType;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
@@ -92,7 +92,7 @@ class LocalOutlierFactorTest extends TestCase
     {
         $this->assertInstanceOf(LocalOutlierFactor::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
-        $this->assertInstanceOf(Ranking::class, $this->estimator);
+        $this->assertInstanceOf(Scoring::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
     }